orphan snapshots cleaning
- deleting snapshots at cli launch (with confirmation)
- option to skip orphan snapshots checking
- code separation
- more tests
